Move ConditionsEditor from src/components/stepforms to src/components
