+++ .storybook/bundle.ts
@@ -8,8 +8,8 @@ import RenameStepForm from '../src/components/stepforms/RenameStepForm.vue';
 import Pipeline from '../src/components/Pipeline.vue';
 import ResizablePanels from '../src/components/ResizablePanels.vue';
 import Step from '../src/components/Step.vue';
-import ConditionForm from '../src/components/stepforms/ConditionsEditor/ConditionForm.vue';
-import ConditionsEditor from '../src/components/stepforms/ConditionsEditor/ConditionsEditor.vue';
+import ConditionForm from '../src/components/ConditionsEditor/ConditionForm.vue';
+import ConditionsEditor from '../src/components/ConditionsEditor/ConditionsEditor.vue';
 
 export { ConditionForm, ConditionsEditor, DataViewer, RenameStepForm, Pipeline, ResizablePanels, Step };
 export { setupStore } from '../src/store';
