+++ src/components/ConditionsEditor/ConditionForm.vue
