+++ src/components/ConditionsEditor/ConditionsEditor.vue
